pyproject.toml
[tool.poetry]
name = "rbot"
version = "0.1.0"
description = ""
authors = ["lap1dem <vadym.bidula@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
pillow = "^9.3.0"
py-cord = "^2.3.2"
redis = "^4.4.0"
requests = "^2.28.1"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
